chore: extract dataSet paylod import in a fixture file
